% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_groups.R
\name{add_groups}
\alias{add_groups}
\title{Add groups based on experiment name.}
\usage{
add_groups(x, group_ids = list(All_samples = "All samples"))
}
\arguments{
\item{x}{A dataframe containing a column named Experiments, based on which you will define the groups.}

\item{group_ids}{List containing group name as name of the list element, and an identifier unique for this group as list element.}
}
\description{
Add group identifies based on strings included in the Experiments name.
}
