% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments}
\alias{experiments}
\title{View names of the different experiments}
\usage{
experiments(x)
}
\arguments{
\item{x}{A dataframe.}
}
\description{
Prints a list of the different experiment names
}
\examples{
experiments(my_data)

}
