% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_Andromeda_score.R
\name{MQ_evidence_Andromeda_score}
\alias{MQ_evidence_Andromeda_score}
\title{Plot Andromeda score distribution using MaxQuant output}
\usage{
MQ_evidence_Andromeda_score(x)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}
}
\value{
Returns a density plot.
}
\description{
Uses the output from MaxQuant (evidence.txt) to plot Andromeda score distribution for the identified experiments.
}
