% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_charge_state.R
\name{MQ_evidence_charge_state}
\alias{MQ_evidence_charge_state}
\title{Charge state distribution using MaxQuant output}
\usage{
MQ_evidence_charge_state(x, per_sample = FALSE, ncol = 5)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}

\item{per_sample}{Logical with default value equal to FALSE. By default a summary from all experiments will
be plotted. If set to TRUE, individual experiments will be shown in different graphs.}

\item{ncol}{Number of columns when plotting individual experiments. Default value is 5.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to create a bargraph of
charge state distribution as a total or splitted by experiment.
}
\examples{
# plot charge state as a total
MQ_evidence_charge_state(filtered_evidence_output, per_sample = TRUE, ncol = 6)

#plot charge state per experiment
MQ_evidence_charge_state(filtered_evidence_output, per_sample = TRUE, ncol = 6)

}
