% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_intensity_boxplot.R
\name{MQ_evidence_intensity_boxplot}
\alias{MQ_evidence_intensity_boxplot}
\title{Plot peptide intensities for individual experiments using boxplots}
\usage{
MQ_evidence_intensity_boxplot(x, color_groups = FALSE)
}
\arguments{
\item{x}{The evidence.txt file from MaxQuant output.}

\item{color_groups}{Logical. Is True, colors will be assigned to the boxplots based on the
my_groups column. Use add_groups function first to create the 'add_groups' column. Default
value is FALSE, which will give boxplots with the same color.}
}
\value{
Boxplots of log10 peptidide intensity.
}
\description{
Make boxplots for peptide intensity for the individual experiments. Optionally
you can color them by their group_id (see color_groups argument)
}
