% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_missed_cleavages.R
\name{MQ_evidence_missed_cleavages}
\alias{MQ_evidence_missed_cleavages}
\title{Plot number of missed cleavages using MaxQuant output}
\usage{
MQ_evidence_missed_cleavages(x, per_sample = FALSE, ncol = 5)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}

\item{per_sample}{Logical with default value equal to FALSE. By default a summary from all experiments will
be plotted. If set to TRUE, individual experiments will be shown in different graphs.}

\item{ncol}{Number of columns when plotting individual experiments. Default value is 5.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to create a bargraph of
the number of missed cleavages as a total or splitted by experiment.
}
\examples{
# plot number of missed cleavages as a total
MQ_evidence_missed_clevages(filtered_evidence_output, per_sample = TRUE, ncol = 6)

#plot number of missed cleavages per experiment
MQ_evidence_missed_clevages(filtered_evidence_output, per_sample = TRUE, ncol = 6)

}
