% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_MSMS_count.R
\name{MQ_evidence_MSMS_count}
\alias{MQ_evidence_MSMS_count}
\title{Plot MS/MS count distribution using MaxQuant output}
\usage{
MQ_evidence_MSMS_count(x)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to plot MS/MS count distribution for the identified experiments.
}
