% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_pept_int.R
\name{MQ_evidence_pept_int}
\alias{MQ_evidence_pept_int}
\title{Plot peptide intensity distrubition using MaxQuant output}
\usage{
MQ_evidence_pept_int(x)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to create a density plot of
peptide resolution distribution for the individual experiments.
}
