% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_ppm_error.R
\name{MQ_evidence_ppm_error}
\alias{MQ_evidence_ppm_error}
\title{Plot ppm error distribution using MaxQuant output}
\usage{
MQ_evidence_ppm_error(x)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to plot ppm error distribution for the identified peptides.
}
