% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_resolution.R
\name{MQ_evidence_resolution}
\alias{MQ_evidence_resolution}
\title{Plot instrument resolution distribution using MaxQuant output}
\usage{
MQ_evidence_resolution(x)
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}
}
\description{
Uses the output from MaxQuant (evidence.txt) to plot resolution distribution in the individual experiments.
}
