% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_evidence_RT.R
\name{MQ_evidence_RT}
\alias{MQ_evidence_RT}
\title{Plot run retention time distribution using MaxQuant output}
\usage{
MQ_evidence_RT(x, xlim = c(0, 60))
}
\arguments{
\item{x}{The filtered 'evidence.txt' file. Use MQ_remove_CR first to remove contaminants and reverse hits.}

\item{xlim}{Numeric vector containing two elements defining the min and max limits of the x axis. Default value is xlim = c(0, 60).}
}
\description{
Uses the output from MaxQuant (evidence.txt) to plot retention time distribution for the individual experiments.
}
