% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_remove_bysite.R
\name{MQ_remove_bysite}
\alias{MQ_remove_bysite}
\title{Remove proteins indentified only by modified peptides from MaxQuant output files}
\usage{
MQ_remove_bysite(x)
}
\arguments{
\item{x}{Input dataframe.}
}
\value{
Returns a filtered dataframe similar to the imput.
}
\description{
Removes proteins identified only by modified peptides (Only by site)
from the MaxQuant output files.
}
\examples{
filtered_data <- MQ_remove_bysite(raw_data)

}
