% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_remove_CR.R
\name{MQ_remove_CR}
\alias{MQ_remove_CR}
\title{Remove reverse hits and contaminants from MaxQuant output files}
\usage{
MQ_remove_CR(data)
}
\arguments{
\item{data}{Input dataframe.}
}
\value{
Returns a filtered dataframe similar to the imput.
}
\description{
Removes reverse hits and contamintants from the MaxQuant output files.
}
