% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_remove_reverse.R
\name{MQ_remove_reverse}
\alias{MQ_remove_reverse}
\title{Remove reverse sequences from MaxQuant output files}
\usage{
MQ_remove_reverse(x)
}
\arguments{
\item{x}{Input dataframe.}
}
\value{
Returns a filtered dataframe similar to the imput.
}
\description{
Removes reverse sequences from the MaxQuant output files.
}
\examples{
filtered_data <- MQ_remove_reverse(raw_data)

}
