% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_summary_MS1_spectra.R
\name{MQ_summary_MS1_spectra}
\alias{MQ_summary_MS1_spectra}
\title{Number of MS1 spectra aquired (read from MaxQuant summary.txt)}
\usage{
MQ_summary_MS1_spectra(x, grouping = FALSE)
}
\arguments{
\item{x}{The processed summary.txt output file}

\item{grouping}{Default value is FALSE. If set to true, the different experiments
will be colored based on the `my_groups` column. Use \code{\link{add_groups}} function
to define the groups based on experiment names.}
}
\description{
Reads MaxQuant summary.txt output file and plots the number of
   MS1 spectra aquired. Use the `add_groups` function first to add groups based
   on the experiment name, and color the different raw files based on the groups.
}
