% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_summary_peaks_seq_percent.R
\name{MQ_summary_peaks_seq_percent}
\alias{MQ_summary_peaks_seq_percent}
\title{Percent of peaks sequenced (read from MaxQuant summary.txt)}
\usage{
MQ_summary_peaks_seq_percent(x, grouping = FALSE)
}
\arguments{
\item{x}{The processed summary.txt output file}

\item{grouping}{Default value is FALSE. If set to true, the different experiments
will be colored based on the `my_groups` column. Use \code{\link{add_groups}} function
to define the groups based on experiment names.}
}
\description{
Reads MaxQuant summary.txt output file and plots the percentage of
   peaks sequenced. Use the `add_groups` function first to add groups based
   on the experiment name, and color the different raw files based on the groups.
}
