% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_summary_pept_id.R
\name{MQ_summary_pept_id}
\alias{MQ_summary_pept_id}
\title{Number of peptide sequences identified (read from MaxQuant summary.txt)}
\usage{
MQ_summary_pept_id(x, grouping = FALSE)
}
\arguments{
\item{x}{The processed summary.txt output file}

\item{grouping}{Default value is FALSE. If set to true, the different experiments
will be colored based on the `my_groups` column. Use \code{\link{add_groups}} function
to define the groups based on experiment names.}
}
\description{
Reads MaxQuant summary.txt output file and plots the number of
   peptide sequences identified Use the `add_groups` function first to add groups based
   on the experiment name, and color the different raw files based on the groups.
}
