% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_groups.R
\name{add_groups}
\alias{add_groups}
\title{Add groups based on experiment name.}
\usage{
add_groups(x, group_ids = list(All_samples = "All samples"))
}
\arguments{
\item{x}{A dataframe containing a column named Experiments, based on which you will define the groups.}

\item{group_ids}{List containing group name as name of the list element, and an identifier unique for this group as list element.}
}
\description{
Add group identifies based on strings included in the Experiments name.
}
\examples{
   # first define the group_ids list
   group_ids <- list(GroupA = "group_A",
                 GroupB = "group_B")
   # 2nd assign different experiments into groups. Output is saved in the 'my_groups' column
   data2 <- add_groups(data, group_ids)

}
