% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_summary_MSvsMSMS.R
\name{MQ_summary_MSvsMSMS}
\alias{MQ_summary_MSvsMSMS}
\title{View number of MS1 and MS2 spectra per sample from summary.txt file}
\usage{
MQ_summary_MSvsMSMS(x, lab_size = 8)
}
\arguments{
\item{x}{The summary.txt file from MaxQuant output.}

\item{lab_size}{Size of the text labels.}
}
\value{
A barplot
}
\description{
Make a side-by-side barplot with the number of MS1 and the
number of MS2 scans per sample.
}
