% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_summary_MSvsMSMS.R
\name{MQ_summary_MSvsMSMS}
\alias{MQ_summary_MSvsMSMS}
\title{View number of MS1 and MS2 spectra per sample from summary.txt file}
\usage{
MQ_summary_MSvsMSMS(x, num_size = 4, num_angle = 0, num_vjust = 0.5,
  num_hjust = 1)
}
\arguments{
\item{x}{The summary.txt file from MaxQuant output.}

\item{num_size}{Numeric label size. Default value is 4.}

\item{num_angle}{Numeric label label angle. Default value is 0.}

\item{num_vjust}{Vertical justification for numeric label. Default value is 0.5.}

\item{num_hjust}{Horizontal justification for numeric label. Default value is 1.}
}
\value{
A barplot
}
\description{
Make a side-by-side barplot with the number of MS1 and the
number of MS2 scans per sample.
}
